version 0.2.4
bug fixes regarding to the new attendance tracking feature.
previously when we hit the route every student got skips because the payload did not have any phone number to send sms to.
now we are calling students info from the mongodb using the _id coming from the payload the the attendance route.